feat(options.configurableGlobals): allow shims to mutate globals
This feature fixes a TODO that said the fast freezing of globals
that are generally "stable" is for performance, but isn't totally
correct.

I've run into problems trying to provide a wrapper for all globals
from within a shim.  Notably, it was impossible to replace Array,
Object, and other functions in the global namespace because they
are frozen.